Rearrange types to split up monolithic declaration
