+++ src/components/CriteriaCardUnknown.test.tsx
@@ -1,14 +1,8 @@
-import { render, screen } from '@testing-library/react';
+import { render } from '@testing-library/react';
 import arrayMutators from 'final-form-arrays';
 import React from 'react';
 import { Form } from 'react-final-form';
-import {
-  CriteriaCardGroupType,
-  CriteriaCardTerminalType,
-  FormValues,
-} from '../form/types';
 import withIntlConfiguration from '../test/util/withIntlConfiguration';
-import CriteriaCardToolbox from './CriteriaCardToolbox';
 import CriteriaCard from './CriteriaCard';
 
 const noop = () => ({});
