+++ src/components/CriteriaPatronGroup.test.tsx
@@ -3,7 +3,7 @@ import userEvent from '@testing-library/user-event';
 import arrayMutators from 'final-form-arrays';
 import React from 'react';
 import { Form } from 'react-final-form';
-import { FormValues } from '../form/types';
+import FormValues from '../types/FormValues';
 import withIntlConfiguration from '../test/util/withIntlConfiguration';
 import CriteriaCard from './CriteriaCard';
 import { QueryClient, QueryClientProvider } from 'react-query';
