+++ src/form/sections/SchedulingMenu.test.tsx
@@ -2,8 +2,9 @@ import { render, screen } from '@testing-library/react';
 import React from 'react';
 import { Form } from 'react-final-form';
 import withIntlConfiguration from '../../test/util/withIntlConfiguration';
+import FormValues from '../../types/FormValues';
+import SchedulingFrequency from '../../types/SchedulingFrequency';
 import SchedulingMenu, { getIntervalLabel } from './SchedulingMenu';
-import { FormValues, SchedulingFrequency } from '../types';
 
 const noop = () => ({});
