+++ src/types/ConvenientConstants.ts
@@ -1,0 +1,7 @@
+export default {
+  NEWLINE: '\n',
+  NEWLINE_MICROSOFT: '\r\n',
+  TAB: '\t',
+  COMMA: ',',
+  SPACE: ' ',
+};
