+++ src/types/LengthControl.ts
@@ -1,0 +1,6 @@
+export default interface LengthControl {
+  character: string;
+  length: number;
+  direction: 'LEFT' | 'RIGHT';
+  truncate: boolean;
+}
