+++ src/types/SchedulingFrequency.ts
@@ -1,0 +1,8 @@
+enum SchedulingFrequency {
+  Manual = 'NONE',
+  Hours = 'HOUR',
+  Days = 'DAY',
+  Weeks = 'WEEK',
+}
+
+export default SchedulingFrequency;
